adding testing to some methods in repo also adding swipe to delete in fav make the settings list relevant to the real one fix some bugs related to connections fix some bugs in the weather response model make it hard to get the data

+++ app/src/main/java/inc/moe/weather/model/Hourly.kt
@@ -21,5 +21,5 @@ data class Hourly(
     val wind_deg: Long,
     val wind_gust: Double,
     val weather: List<Weather>,
-    val pop: Long,
+    val pop: Double,
 ): Serializable
